move registry cache ensurer into own package
