+++ pkg/webhook/controlplane/registrycache/registry_cache_test.go
@@ -1,13 +1,9 @@
-package controlplane
+package registrycache
 
 import (
-	"context"
 	"encoding/base64"
 
-	extensionscontroller "github.com/gardener/gardener/extensions/pkg/controller"
-	gcontext "github.com/gardener/gardener/extensions/pkg/webhook/context"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
-	"github.com/go-logr/logr"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
@@ -16,19 +12,11 @@ import (
 
 var _ = Describe("registry cache files", func() {
 	var (
-		ctx  context.Context
-		gctx gcontext.GardenContext
-		log  logr.Logger
-
 		regCaches []config.RegistryCacheConfiguration
 		files     []extensionsv1alpha1.File
 	)
 
 	BeforeEach(func() {
-		ctx = context.Background()
-		gctx = &fakeGardenContext{}
-		log = logr.Discard()
-
 		regCaches = []config.RegistryCacheConfiguration{{
 			Server:       "https://foo.com",
 			Cache:        "https://foo-cache.com",
@@ -38,18 +26,20 @@ var _ = Describe("registry cache files", func() {
 	})
 
 	It("should not add anything with an empty list (AdditionalProvisionFiles)", func() {
-		e := NewEnsurer(nil, log)
+		e := &Ensurer{}
 
-		Expect(e.EnsureAdditionalProvisionFiles(ctx, gctx, &files, nil)).To(Succeed())
+		Expect(e.EnsureCaches(&files)).To(Succeed())
 		Expect(files).NotTo(ContainElement(
 			HaveField("Path", ContainSubstring("hosts.toml")),
 		))
 	})
 
 	It("should inject the config without CA (AdditionalProvisionFiles)", func() {
-		e := NewEnsurer(regCaches, log)
+		e := &Ensurer{
+			Caches: regCaches,
+		}
 
-		Expect(e.EnsureAdditionalProvisionFiles(ctx, gctx, &files, nil)).To(Succeed())
+		Expect(e.EnsureCaches(&files)).To(Succeed())
 
 		expetedData := `# Created by gardener-extension-provider-stackit
 server = 'https://foo.com'
@@ -71,9 +61,11 @@ capabilities = ['pull']
 		dummyCA := []byte("--- Certificate_---\nfoo\nbar")
 		regCaches[0].CABundle = dummyCA
 		encoded := base64.StdEncoding.EncodeToString(dummyCA)
-		e := NewEnsurer(regCaches, log)
+		e := &Ensurer{
+			Caches: regCaches,
+		}
 
-		Expect(e.EnsureAdditionalFiles(ctx, gctx, &files, nil)).To(Succeed())
+		Expect(e.EnsureCaches(&files)).To(Succeed())
 
 		Expect(files).To(ContainElements(
 			And(
@@ -90,10 +82,3 @@ capabilities = ['pull']
 		))
 	})
 })
-
-type fakeGardenContext struct {
-}
-
-func (f *fakeGardenContext) GetCluster(ctx context.Context) (*extensionscontroller.Cluster, error) {
-	return &extensionscontroller.Cluster{}, nil
-}
